requirements.txt
asgiref==3.2.3
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-extensions==0.6.0
nltk==3.4.5
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
textblob==0.15.3
